Basic Level Transitions
Now there is basic transitions between levels. When the player has completed a level (killed all enemies) and moves past the right edge of the screen, the level will increase. If they move to the left of the screen, the level will decrease (BUG: unfortunately this means the player teleports to right side of screen on level 1).
- Added MoveLevel method to Player, which queues the appropriate level increase/decrease command
- Added level transition logic to LevelManager to use player's MoveLevel method and update the player's position to opposite side of screen.
- Added logic to InitializeEntities in GameManager so that the player is consistent between level transitions instead of respawning.